+++ src/components/playlist/Playlist.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import { Component } from 'react'
 import { Outlet } from 'react-router-dom'
 
 import Tab from 'components/generics/Tab'
@@ -10,14 +10,19 @@ export default class Playlist extends Component {
             <div id="playlist" className="box">
                 <nav className="tab-bar">
                     <Tab
+                        to="/playlist/queueing"
+                        iconName="chevron-right"
+                        name="Queuing"
+                    />
+                    <Tab
                         to="/playlist/played"
                         iconName="chevron-left"
                         name="Played"
                     />
                     <Tab
-                        to="/playlist/queueing"
-                        iconName="chevron-right"
-                        name="Queuing"
+                        to="/playlist/player-errors"
+                        iconName="exclamation-triangle"
+                        name="Errors"
                     />
                 </nav>
                 <Outlet />