+++ src/components/user/User.jsx
@@ -1,10 +1,10 @@
 import PropTypes from 'prop-types'
-import React, { Component } from 'react'
+import { Component } from 'react'
 import { connect } from 'react-redux'
 
 import { FormBlock, InputField } from 'components/generics/Form'
-import { permissionLevels } from 'components/permissions/Users'
 import { userPropType } from 'serverPropTypes/users'
+import { permissionLevels } from 'utils/permissions';
 
 class User extends Component {
     static propTypes = {